+++ MapDriveInator/zips/templates/MapDriveInator/MDI styles template.mm
@@ -13,7 +13,7 @@
             <attribute_contains_condition ATTRIBUTE="modifiedFile" VALUE="true"/>
         </conditional_style>
     </conditional_styles>
-    <properties show_icon_for_attributes="true" show_notes_in_map="false" BookmarksKeys="{&quot;103&quot;:&quot;ID_1479759080&quot;,&quot;118&quot;:&quot;ID_128687180&quot;,&quot;119&quot;:&quot;ID_1586584839&quot;,&quot;97&quot;:&quot;ID_234882181&quot;}" mapVariable_de_prueba="valor variable 75" show_note_icons="true" fit_to_viewport="false;" mapUsesOwnSaveOptions="true" save_folding="default" save_last_visited_node="default" save_modification_times="default"/>
+    <properties show_icon_for_attributes="true" show_notes_in_map="false" show_note_icons="true" fit_to_viewport="false;" MDI_template="v0.0.13"/>
 
 <map_styles>
 <stylenode LOCALIZED_TEXT="styles.root_node" ID="ID_1625586047" STYLE="oval" UNIFORM_SHAPE="true" VGAP_QUANTITY="24 pt">